backend/requirements.txt
fastapi
python-multipart
uvicorn
sqlalchemy
alembic
psycopg2-binary

python-dotenv
openai>=1.0.0
razorpay
brevo_python
passlib[bcrypt]
bcrypt==4.0.1
PyJWT
uuid

pypdf
fpdf

aiohttp
aiodns
websockets

google-cloud-storage